move mm3 hash to blender_utils.py
fixes #125
